Remove proxyquire: enable esm tests
