+++ packages/tools/devtools/devtools-browser-extension/package.json
@@ -44,7 +44,7 @@
 		"test:jest:verbose": "cross-env FLUID_TEST_VERBOSE=1 jest",
 		"test:mocha": "npm run test:mocha:esm && npm run test:mocha:cjs",
 		"test:mocha:cjs": "mocha --recursive \"dist/tsc/test/**/*.test.*js\" --exit",
-		"test:mocha:esm": "echo \"Skipping ESM tests -- See AB#7470\"",
+		"test:mocha:esm": "mocha --recursive \"lib/tsc/test/**/*.test.*js\" --exit",
 		"test:mocha:verbose": "cross-env FLUID_TEST_VERBOSE=1 npm run mocha -- --verbose",
 		"test:unit": "npm run test:mocha",
 		"test:verbose": "concurrently npm:test:mocha:verbose npm:test:jest:verbose",
@@ -136,7 +136,6 @@
 		"jsdom-global": "^3.0.2",
 		"mocha": "^10.2.0",
 		"prettier": "~3.0.3",
-		"proxyquire": "^2.1.3",
 		"puppeteer": "^23.6.0",
 		"rimraf": "^4.4.0",
 		"sinon": "^18.0.1",
