+++ packages/tools/devtools/devtools-browser-extension/src/content/ContentScript.ts
@@ -3,25 +3,9 @@
  * Licensed under the MIT License.
  */
 
-import {
-	type ISourcedDevtoolsMessage,
-	devtoolsMessageSource,
-	isDevtoolsMessage,
-} from "@fluidframework/devtools-core/internal";
-
 import { browser, window } from "../Globals.js";
-import {
-	extensionPopupMessageSource,
-	extensionViewMessageSource,
-	relayMessageToPort,
-} from "../messaging/index.js";
 
-import {
-	contentScriptMessageLoggingOptions,
-	formatContentScriptMessageForLogging,
-} from "./Logging.js";
-
-type Port = chrome.runtime.Port;
+import { runContentScript } from "./ContentScriptContent.js";
 
 /**
  * This module is the extension's Content Script.
@@ -37,67 +21,5 @@ type Port = chrome.runtime.Port;
  * see {@link https://developer.chrome.com/docs/extensions/mv3/devtools/#content-script-to-devtools | here}.
  */
 
-console.log(formatContentScriptMessageForLogging("Initializing Content Script."));
-
-// `window` should always be defined in the Content script context.
-if (window === undefined) {
-	throw new Error("Window object is not defined.");
-}
-
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-
-// Only establish messaging when activated by the Background Worker.
-browser.runtime.onConnect.addListener((backgroundPort: Port) => {
-	console.log(
-		formatContentScriptMessageForLogging("Connection added from Background Worker."),
-	);
-	const allowedMessageSources = new Set([
-		extensionViewMessageSource,
-		extensionPopupMessageSource,
-	]);
-	/**
-	 * Relay messages if they conform to our expected format.
-	 */
-	function relayMessageFromPageToBackground(
-		event: MessageEvent<Partial<ISourcedDevtoolsMessage>>,
-	): void {
-		const message = event.data;
-
-		// Only relay message if it is one of ours, and if the source is the window's Devtools instance
-		// (and not a message originating from the extension).
-		if (isDevtoolsMessage(message) && message.source === devtoolsMessageSource) {
-			relayMessageToPort(
-				message,
-				"webpage",
-				backgroundPort,
-				contentScriptMessageLoggingOptions,
-			);
-		}
-	}
-
-	// Relay messages to the Background Worker as appropriate.
-	window!.addEventListener("message", relayMessageFromPageToBackground);
-
-	// Relay messages from the Background Worker to the inspected window.
-	backgroundPort.onMessage.addListener((message: Partial<ISourcedDevtoolsMessage>) => {
-		// Only relay message if it is one of ours, and if the source is the extension
-		// (and not the window).
-		if (isDevtoolsMessage(message) && allowedMessageSources.has(message.source)) {
-			console.debug(
-				formatContentScriptMessageForLogging(
-					`Relaying message from Background Script to the window:`,
-				),
-				message,
-			);
-			window!.postMessage(message, "*");
-		}
-	});
-
-	// When the extension disconnects, clean up listeners.
-	backgroundPort.onDisconnect.addListener(() => {
-		// Unbind window listener
-		window!.removeEventListener("message", relayMessageFromPageToBackground);
-	});
-});
-
-/* eslint-enable @typescript-eslint/no-non-null-assertion */
+// Tests assume this just calls "runContentScript": any changes to the body of this module should be made to "run", or may require updates to the tests.
+runContentScript(browser, window);
