+++ packages/tools/devtools/devtools-browser-extension/src/test/DevtoolsScript.test.ts
@@ -13,9 +13,8 @@ import { runDevtoolsScript } from "../devtools/DevtoolsScriptContent.js";
 import { stubGlobals } from "./Utilities.js";
 
 describe("Devtools Script unit tests", () => {
-	const sandbox = createSandbox();
-
 	it("Creates the view on load", async () => {
+		const sandbox = createSandbox();
 		const { browser } = stubGlobals();
 		const createSpy = sandbox.spy(browser.devtools.panels, "create");
 		runDevtoolsScript(browser);
